filtros e alteraçao componentes
