+++ src/Components/Carrinho.js
@@ -1,20 +1,27 @@
-import React from "react";
-import styled from "styled-components";
+// import React from "react";
+// import styled from "styled-components";
+// import ItemCarrinho from "./ItemCarrinho";
 
-const ContainerCarrinho = styled.div`
-display: flex;
-flex-direction: column;
-border: 1px solid black;
-height: 100vh;
-`
-export default class Carrinho extends React.Component {
+// const ContainerCarrinho = styled.div`
+// display: flex;
+// flex-direction: column;
+// border: 1px solid black;
+// height: 100vh;
+// `
+// export default class Carrinho extends React.Component {
+//   state = {
+//     carrinho: [],
+//     valorTotal: 0
+//   }
 
-    render() {
-      return (
-        <ContainerCarrinho>
-            <h2>Carrinho:</h2>
-            <p>Valor Total: R$</p>
-        </ContainerCarrinho>
-      );
-    }
-  }
+
+//     render() {
+//       return (
+//         <ContainerCarrinho>
+//             <h2>Carrinho:</h2>
+//             <ItemCarrinho/>
+//             <p>Valor Total: R$</p>
+//         </ContainerCarrinho>
+//       );
+//     }
+//   }
