+++ src/Components/ItemCarrinho.js
@@ -1,0 +1,24 @@
+// import React from "react";
+// import styled from "styled-components";
+
+
+// const CarrinhoItem = styled.div`
+// display: flex;
+// justify-content: space-between;
+// align-items: center;
+// `
+
+
+// export default class ItemCarrinho extends React.Component {
+    
+//     render() {
+//         return (
+//             <CarrinhoItem>
+//                 <p>1x</p>
+//                 <p>Produto 1</p>
+//                 <button> Remover </button>
+//             </CarrinhoItem>
+//         );
+//     }
+// }
+
